pyproject.toml
[project]
name = "fastapi-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "bcrypt==4.0.1",
    "fastapi>=0.115.12",
    "fastapi-mail>=1.6.2",
    "openai>=2.28.0",
    "passlib[bcrypt]>=1.7.4",
    "pip>=26.0.1",
    "prisma>=0.15.0",
    "pydantic>=2.11.4",
    "python-jose>=3.5.0",
    "redis>=7.4.0",
    "requests>=2.32.3",
    "rq>=2.7.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.34.2",
]
[tool.fastapi]
entrypoint = "main:app"
[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
